Filtered data
The data was filtered to only contain values from Canada and USA, and not Nigeria, India, and Germany, since we are only concerned with comparing those two countries. The data was also filtered to remove the  hospital beds per 100 column, since we did not ask for this in our research proposal questions.