Created the code for plot 1
Plot 1 describes life expectancy in Canada vs USA every 5 years from 1990 to 2020. The data had to be filtered even more to be every 5 years and also Canada and USA split up. The reason it is every 5 years is because every year would create a very messy graph that would not give any information to the reader.